mysql/sqlite: move common code to sql/common.go
