+++ stores/sql/sqlite/metrics.go
@@ -33,17 +33,5 @@ func (b *MetricsDatabase) Version(_ context.Context) (string, string, error) {
 }
 
 func (b *MetricsDatabase) Migrate() error {
-	dbIdentifier := "metrics"
-	return performMigrations(b.db, dbIdentifier, []migration{
-		{
-			ID:      "00001_init",
-			Migrate: func(tx sql.Tx) error { return sql.ErrRunV072 },
-		},
-		{
-			ID: "00001_idx_contracts_fcid_timestamp",
-			Migrate: func(tx sql.Tx) error {
-				return performMigration(tx, dbIdentifier, "00001_idx_contracts_fcid_timestamp", b.log)
-			},
-		},
-	}, b.log)
+	return performMigrations(b.db, "metrics", sql.MetricsMigrations(migrationsFs, b.log), b.log)
 }
